requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
libmagic
python-magic
python-magic-bin
langchain_experimental
matplotlib
seaborn
pandasai
selenium
unstructured
streamlit-option-menu
